speedLimit

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/turret/Turret.java
@@ -38,7 +38,7 @@ public class Turret extends SubsystemBase {
                 , ticksSetpoint, TurretConstants.alignEpsilon);
     }
 
-    public void modify(double ticks) {
+    public void modify(int ticks) {
         killedSetpoint += ticks;
     }
